Move SDK telemetry out of tree namespace
